♻️ refactorに伴う修正とprettier
