+++ src/routes/index.svelte
@@ -1,25 +1,23 @@
 <script lang="ts" context="module">
-  import type { Load } from "@sveltejs/kit";
-  import { base as b } from '$app/paths';
+	import type { Load } from '@sveltejs/kit';
+	import { base as b } from '$app/paths';
 
-  export const load: Load = async ({ fetch }) => {
-    const res = await fetch(`${b}/getDocuments`);
-    const json = await res.json();
-    return {
-      props: {
-        articles: json,
-      }
-    };
-  };
+	export const load: Load = async ({ fetch }) => {
+		const res = await fetch(`${b}/list`);
+		const json = await res.json();
+		return {
+			props: { articles: json }
+		};
+	};
 </script>
 
 <script lang="ts">
-  import { base } from '$app/paths';
-  export let articles: { title: string; href: string; }[];
+	import { base } from '$app/paths';
+	export let articles: { title: string; href: string }[];
 </script>
 
 <main>
-  {#each articles as article}
-    <h3><a href="{base}/{article.href}">{article.title}</a></h3>
-  {/each}
+	{#each articles as article}
+		<h3><a href="{base}/{article.href}">{article.title}</a></h3>
+	{/each}
 </main>
